+++ controllers/usersCtrl/current.js
@@ -1,0 +1,9 @@
+const current = async (req, res, next) => {
+    const{ email, subscription} = req.user
+    res.json({
+        email,
+        subscription
+    })
+}
+
+module.exports = current;
