+++ controllers/usersCtrl/index.js
@@ -1,0 +1,15 @@
+const signup = require('./signup');
+const login = require('./login');
+const logout = require('./logout');
+const current = require('./current');
+const subscription = require('./subscription');
+const avatars = require('./avatars');
+
+module.exports = {
+    signup,
+    login,
+    logout,
+    current,
+    subscription,
+    avatars
+};
