+++ controllers/usersCtrl/logout.js
@@ -1,0 +1,8 @@
+const {User} = require("../../models/user")
+const logout = async (req, res, next) => {
+    const { _id } = req.user;
+    await User.findByIdAndUpdate(_id, { token: '' });
+    res.status(204).send()
+}
+
+module.exports = logout;
